pathtrainschedule: Fixup package naming.
This commit fixes up some of the naming in path train schedule to
standardize it with the rest of the repo.
