+++ apps/apps.go
@@ -43,6 +43,7 @@ import (
 	"tidbyt.dev/community/apps/nycbus"
 	"tidbyt.dev/community/apps/ohhighwaysigns"
 	"tidbyt.dev/community/apps/pagerduty"
+	"tidbyt.dev/community/apps/pathtrainschedule"
 	"tidbyt.dev/community/apps/phaseofmoon"
 	"tidbyt.dev/community/apps/pokedex"
 	"tidbyt.dev/community/apps/sbbtimetable"
@@ -95,6 +96,7 @@ func GetManifests() []manifest.Manifest {
 		nycbus.New(),
 		ohhighwaysigns.New(),
 		pagerduty.New(),
+		pathtrainschedule.New(),
 		phaseofmoon.New(),
 		pokedex.New(),
 		sbbtimetable.New(),
@@ -105,7 +107,7 @@ func GetManifests() []manifest.Manifest {
 		twitterfollows.New(),
 		verticalmessage.New(),
 		weathermap.New(),
- 		worldclock.New(),
+		worldclock.New(),
 	}
 }
